+++ app/core/app_state.py
@@ -1,15 +1,28 @@
 import logging
 import os
+from functools import lru_cache
 
+import cachetools
 import yaml
-from fastapi import Request
+from fastapi import BackgroundTasks, Request
+from groundlight import Groundlight
+from model import (
+    Detector,
+)
 
 from .configs import EdgeInferenceConfig, RootEdgeConfig
 from .database import DatabaseManager
 from .edge_inference import EdgeInferenceManager
 from .file_paths import DEFAULT_EDGE_CONFIG_PATH
+from .utils import safe_call_sdk
 
 logger = logging.getLogger(__name__)
+
+
+MAX_SDK_INSTANCES_CACHE_SIZE = 1000
+MAX_DETECTOR_IDS_TTL_CACHE_SIZE = 1000
+TTL_TIME = 600  # 10 minutes
+REFRESH_METADATA_INTERVAL = 30  # in seconds
 
 
 def load_edge_config() -> RootEdgeConfig:
@@ -70,6 +83,63 @@ def get_detector_inference_configs(
         }
 
     return detector_to_inference_config
+
+
+@lru_cache(maxsize=MAX_SDK_INSTANCES_CACHE_SIZE)
+def _get_groundlight_sdk_instance_internal(api_token: str):
+    return Groundlight(api_token=api_token)
+
+
+def get_groundlight_sdk_instance(request: Request):
+    """
+    Returns a (cached) Groundlight SDK instance given an API token.
+    The SDK handles validation of the API token token itself, so there's no
+    need to do that here.
+    """
+    api_token = request.headers.get("x-api-token")
+    return _get_groundlight_sdk_instance_internal(api_token)
+
+
+class TimestampedTTLCache(cachetools.TTLCache):
+    """TTLCache subclass that tracks when items were added to the cache."""
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        self.__timestamps = {}  # Store timestamps for each key
+
+    def __setitem__(self, key, value, cache_setitem=cachetools.Cache.__setitem__):
+        # Track the current time when setting an item
+        self.__timestamps[key] = self.timer()
+        super().__setitem__(key, value, cache_setitem)
+
+    def __delitem__(self, key, cache_delitem=cachetools.Cache.__delitem__):
+        super().__delitem__(key, cache_delitem)
+        self.__timestamps.pop(key, None)
+
+    def get_timestamp(self, key):
+        """Get the timestamp when an item was added to the cache."""
+        return self.__timestamps.get(key)
+
+
+def refresh_detector_metadata_if_needed(detector_id: str, gl: Groundlight, background_tasks: BackgroundTasks) -> None:
+    """Check if detector metadata needs refreshing and schedule a background task if metadata is too old."""
+    cached_value_age = get_detector_metadata.cache.timer() - get_detector_metadata.cache.get_timestamp(detector_id)
+    if cached_value_age > REFRESH_METADATA_INTERVAL:
+        get_detector_metadata.cache.pop(detector_id, None)
+        background_tasks.add_task(get_detector_metadata, detector_id=detector_id, gl=gl)
+
+
+@cachetools.cached(
+    cache=TimestampedTTLCache(maxsize=MAX_DETECTOR_IDS_TTL_CACHE_SIZE, ttl=TTL_TIME),
+    key=lambda detector_id, gl: detector_id,
+)
+def get_detector_metadata(detector_id: str, gl: Groundlight) -> Detector:
+    """
+    Returns detector metadata from the Groundlight API.
+    Caches the result so that we don't have to make an expensive API call every time.
+    """
+    detector = safe_call_sdk(gl.get_detector, id=detector_id)
+    return detector
 
 
 class AppState:
